+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- metadata and screenshots to app store
+
 ## [1.0.0] - 2021-01-06
 
 ### Added